HW 9.1 and HW 10
This commit includes HW9 updated and HW10. The previous starting with HW 8.1  are named incorrectly, I was supposed to type HW9.1 and so on.
In the HW9.1, I fixed the registration and login views. They are using bootstrap now, and  also fixed the error messages so they don't move my textboxes.
Lastly, for HW10 this is what I have so far.